experiment with multimethod for draw transform
