+++ requirements-ci.txt
@@ -13,3 +13,4 @@ dask[delayed]
 autoray
 matplotlib
 opt_einsum
+multimethod
