+++ requirements.txt
@@ -15,3 +15,4 @@ autoray>=0.2.5
 matplotlib==3.4
 black>=21
 opt_einsum==3.3
+multimethod
